+++ nuclia/sdk/kb.py
@@ -1,7 +1,7 @@
 from nucliadb_models.resource import Resource, ResourceList
 
 from nuclia.data import get_auth
-from nuclia.decorators import kb
+from nuclia.decorators import kb, pretty
 from nuclia.lib.kb import NucliaDBClient
 from nuclia.sdk.auth import NucliaAuth
 from nuclia.sdk.search import NucliaSearch
@@ -25,7 +25,15 @@ class NucliaKB:
             print(f"{resource.id} {resource.icon:30} {resource.title}")
 
     @kb
-    def get_resource_by_slug(self, *, ndb: NucliaDBClient, slug: str) -> Resource:
+    @pretty
+    def get_resource_by_id(self, *, ndb: NucliaDBClient, rid: str, **kwargs) -> Resource:
+        return ndb.ndb.get_resource_by_id(
+            kbid=ndb.kbid, rid=rid, query_params={"show": "values"}
+        )
+
+    @kb
+    @pretty
+    def get_resource_by_slug(self, *, ndb: NucliaDBClient, slug: str, **kwargs) -> Resource:
         return ndb.ndb.get_resource_by_slug(
             kbid=ndb.kbid, slug=slug, query_params={"show": "values"}
         )